+++ src/test/java/r11/citrus/s3/config/EndpointConfig.java
@@ -1,7 +1,9 @@
-package r11.citrus.s3;
+package r11.citrus.s3.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import r11.citrus.s3.S3Endpoint;
+import r11.citrus.s3.host.S3AbstractHost;
 
 @Configuration
 public class EndpointConfig {
