+++ src/test/java/r11/citrus/s3/test/AdvancedOperationTests.java
@@ -1,49 +1,33 @@
-package r11.citrus.s3;
+package r11.citrus.s3.test;
 
 import com.consol.citrus.annotations.CitrusTest;
 import com.consol.citrus.dsl.testng.TestNGCitrusTestRunner;
 import com.consol.citrus.validation.binary.BinaryMessageValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.testng.annotations.Test;
+import r11.citrus.s3.S3Endpoint;
+import r11.citrus.s3.S3EndpointResponse;
+import r11.citrus.s3.S3Message;
+import r11.citrus.s3.S3RequestType;
+import r11.citrus.s3.host.S3AbstractHost;
 
 import java.io.IOException;
 
-@Test(testName = "BasicOperation")
-public class BasicOperationTests extends TestNGCitrusTestRunner {
+@Test(testName = "AdvancedOperation")
+public class AdvancedOperationTests extends TestNGCitrusTestRunner {
     @Autowired
     private S3Endpoint s3Endpoint;
-    @Autowired S3AbstractHost s3AbstractHost;
+    @Autowired
+    private S3AbstractHost s3AbstractHost;
 
     private final String bucket = "testbucket1";
     private final String key = "log.txt";
     private final String testValue = "testValue";
 
     @CitrusTest
-    public void createBucketTest() {
+    public void createBucketAndUploadTest() {
         //Put request message
-        S3Message m1 = S3Message.builder().bucket(bucket).method(S3RequestType.CREATE_BUCKET).build();
-
-        //Upload file to S3
-        send(sendMessageBuilder -> sendMessageBuilder
-                .endpoint(s3Endpoint)
-                .message(m1)
-                .payload(testValue)
-        );
-        //Confirm file upload successful
-        receive(receive -> receive
-                .endpoint(s3Endpoint)
-                .payload(S3EndpointResponse.CREATE_BUCKET_SUCCESS)
-        );
-        s3AbstractHost.deleteBucket(bucket);
-
-    }
-
-
-    @CitrusTest
-    public void putFileTest() {
-        s3AbstractHost.createBucket(bucket);
-        //Put request message
-        S3Message m1 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.PUT).build();
+        S3Message m1 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.PUT_BUCKET_CREATE).build();
 
         //Upload file to S3
         send(sendMessageBuilder -> sendMessageBuilder
@@ -58,15 +42,14 @@ public class BasicOperationTests extends TestNGCitrusTestRunner {
         );
 //        s3AbstractHost.deleteObject(bucket, key);
         s3AbstractHost.deleteBucket(bucket);
-
     }
 
     @CitrusTest
-    public void getFileTest() throws IOException {
+    public void getFileAndDeleteTest() throws IOException {
         s3AbstractHost.createBucket(bucket);
         s3AbstractHost.createObject(bucket, key, testValue);
         //Get request message
-        S3Message m2 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.GET).build();
+        S3Message m2 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.GET_DELETE).build();
 
         //Send S3 file request
         send(sendMessageBuilder -> sendMessageBuilder
@@ -81,29 +64,6 @@ public class BasicOperationTests extends TestNGCitrusTestRunner {
         );
 //        s3AbstractHost.deleteObject(bucket, key);
         s3AbstractHost.deleteBucket(bucket);
-
-    }
-
-    @CitrusTest
-    public void deleteFileTest() throws IOException {
-        s3AbstractHost.createBucket(bucket);
-        s3AbstractHost.createObject(bucket, key, testValue);
-        //Delete object request
-        S3Message m3 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.DELETE).build();
-
-        //Send delete object request
-        send(sendMessageBuilder -> sendMessageBuilder
-                .endpoint(s3Endpoint)
-                .message(m3)
-        );
-        //Confirm file delete successful
-        receive(receive -> receive
-                .endpoint(s3Endpoint)
-                .payload(S3EndpointResponse.DELETE_OBJECT_SUCCESS)
-        );
-//        s3AbstractHost.deleteObject(bucket, key);
-        s3AbstractHost.deleteBucket(bucket);
-
     }
 
     @CitrusTest
@@ -125,5 +85,4 @@ public class BasicOperationTests extends TestNGCitrusTestRunner {
         s3AbstractHost.deleteBucket(bucket);
 
     }
-
 }
