+++ src/test/java/r11/citrus/s3/test/BasicOperationTests.java
@@ -1,28 +1,55 @@
-package r11.citrus.s3;
+package r11.citrus.s3.test;
 
 import com.consol.citrus.annotations.CitrusTest;
 import com.consol.citrus.dsl.testng.TestNGCitrusTestRunner;
 import com.consol.citrus.validation.binary.BinaryMessageValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.testng.annotations.Test;
+import r11.citrus.s3.S3Endpoint;
+import r11.citrus.s3.S3EndpointResponse;
+import r11.citrus.s3.S3Message;
+import r11.citrus.s3.S3RequestType;
+import r11.citrus.s3.host.S3AbstractHost;
 
 import java.io.IOException;
 
-@Test(testName = "AdvancedOperation")
-public class AdvancedOperationTests extends TestNGCitrusTestRunner {
+@Test(testName = "BasicOperation")
+public class BasicOperationTests extends TestNGCitrusTestRunner {
     @Autowired
     private S3Endpoint s3Endpoint;
     @Autowired
-    private S3AbstractHost s3AbstractHost;
+    S3AbstractHost s3AbstractHost;
 
     private final String bucket = "testbucket1";
     private final String key = "log.txt";
     private final String testValue = "testValue";
 
     @CitrusTest
-    public void createBucketAndUploadTest() {
+    public void createBucketTest() {
         //Put request message
-        S3Message m1 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.PUT_BUCKET_CREATE).build();
+        S3Message m1 = S3Message.builder().bucket(bucket).method(S3RequestType.CREATE_BUCKET).build();
+
+        //Upload file to S3
+        send(sendMessageBuilder -> sendMessageBuilder
+                .endpoint(s3Endpoint)
+                .message(m1)
+                .payload(testValue)
+        );
+        //Confirm file upload successful
+        receive(receive -> receive
+                .endpoint(s3Endpoint)
+                .payload(S3EndpointResponse.CREATE_BUCKET_SUCCESS)
+        );
+        s3AbstractHost.deleteBucket(bucket);
+
+    }
+
+
+    @CitrusTest
+    public void putFileTest() {
+        s3AbstractHost.createBucket(bucket);
+        //Put request message
+        S3Message m1 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.PUT).build();
 
         //Upload file to S3
         send(sendMessageBuilder -> sendMessageBuilder
@@ -37,14 +64,15 @@ public class AdvancedOperationTests extends TestNGCitrusTestRunner {
         );
 //        s3AbstractHost.deleteObject(bucket, key);
         s3AbstractHost.deleteBucket(bucket);
+
     }
 
     @CitrusTest
-    public void getFileAndDeleteTest() throws IOException {
+    public void getFileTest() throws IOException {
         s3AbstractHost.createBucket(bucket);
         s3AbstractHost.createObject(bucket, key, testValue);
         //Get request message
-        S3Message m2 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.GET_DELETE).build();
+        S3Message m2 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.GET).build();
 
         //Send S3 file request
         send(sendMessageBuilder -> sendMessageBuilder
@@ -59,6 +87,29 @@ public class AdvancedOperationTests extends TestNGCitrusTestRunner {
         );
 //        s3AbstractHost.deleteObject(bucket, key);
         s3AbstractHost.deleteBucket(bucket);
+
+    }
+
+    @CitrusTest
+    public void deleteFileTest() throws IOException {
+        s3AbstractHost.createBucket(bucket);
+        s3AbstractHost.createObject(bucket, key, testValue);
+        //Delete object request
+        S3Message m3 = S3Message.builder().bucket(bucket).key(key).method(S3RequestType.DELETE).build();
+
+        //Send delete object request
+        send(sendMessageBuilder -> sendMessageBuilder
+                .endpoint(s3Endpoint)
+                .message(m3)
+        );
+        //Confirm file delete successful
+        receive(receive -> receive
+                .endpoint(s3Endpoint)
+                .payload(S3EndpointResponse.DELETE_OBJECT_SUCCESS)
+        );
+//        s3AbstractHost.deleteObject(bucket, key);
+        s3AbstractHost.deleteBucket(bucket);
+
     }
 
     @CitrusTest
@@ -80,4 +131,5 @@ public class AdvancedOperationTests extends TestNGCitrusTestRunner {
         s3AbstractHost.deleteBucket(bucket);
 
     }
+
 }
