+++ src/test/resources/citrus-application.properties
@@ -1,2 +1,2 @@
 #Spring Java configuration class holding Citrus endpoint config
-citrus.spring.java.config=r11.citrus.s3.EndpointConfig
+citrus.spring.java.config=r11.citrus.s3.config.EndpointConfig
